+++ examples/base/image_classifier.h
@@ -12,32 +12,21 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#ifndef TNN_EXAMPLES_SAMPLES_TIMER_H_
-#define TNN_EXAMPLES_SAMPLES_TIMER_H_
+#ifndef TNN_EXAMPLES_BASE_IMAGE_CLASSIFIER_H_
+#define TNN_EXAMPLES_BASE_IMAGE_CLASSIFIER_H_
 
-#include <chrono>
+#include <algorithm>
+#include <iostream>
 #include <string>
+#include <vector>
 
-#include "tnn/core/macro.h"
+#include "tnn_sdk_sample.h"
 
-namespace TNN_NS {
-
-using std::chrono::time_point;
-using std::chrono::system_clock;
-
-class SampleTimer {
+class ImageClassifier : public TNN_NS::TNNSDKSample {
 public:
-    SampleTimer() {};
-    void Start();
-    void Stop();
-    void Reset();
-    double GetTime();
-
-private:
-    time_point<system_clock> start_;
-    time_point<system_clock> stop_;
+    ~ImageClassifier();
+    ImageClassifier();
+    int Classify(std::shared_ptr<TNN_NS::Mat> image, int input_width, int input_length, int &class_id);
 };
 
-} // namespace TNN_NS
-
-#endif // TNN_TEST_TIMER_H_ 
+#endif // TNN_EXAMPLES_BASE_IMAGE_CLASSIFIER_H_
