+++ examples/base/sample_timer.cc
@@ -12,7 +12,7 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#include "SampleTimer.h"
+#include "sample_timer.h"
 
 #include <cmath>
